+++ src/hexdoc/cli/app.py
@@ -20,7 +20,12 @@ from hexdoc.__version__ import VERSION
 from hexdoc.core import I18n, ModResourceLoader, ResourceLocation
 from hexdoc.core.properties import AnimationFormat
 from hexdoc.data.metadata import HexdocMetadata
-from hexdoc.data.sitemap import delete_updated_books, dump_sitemap, load_sitemap
+from hexdoc.data.sitemap import (
+    SitemapMarker,
+    delete_updated_books,
+    dump_sitemap,
+    load_sitemap,
+)
 from hexdoc.graphics import DebugType, ModelRenderer
 from hexdoc.graphics.loader import ImageLoader
 from hexdoc.jinja.render import create_jinja_env, get_templates, render_book
@@ -285,6 +290,10 @@ def build(
         env = create_jinja_env(pm, props.template.include, props_file)
 
         logger.info(f"Rendering book for {len(books)} language(s).")
+
+        site_dir.mkdir(parents=True, exist_ok=True)
+        pm.pre_render_site(books, env, site_dir, props.template.include)
+
         for book_info in books:
             try:
                 templates = get_templates(
@@ -337,6 +346,8 @@ def build(
                     raise
                 logger.exception(f"Failed to render book for {book_info.language}")
 
+        pm.post_render_site(books, env, site_dir, props.template.include)
+
     logger.info("Done.")
     return site_dir
 
@@ -375,13 +386,17 @@ def merge(
     root_version: Version | None = None
     root_redirect: str | None = None
 
+    def add_redirect(marker: SitemapMarker, path: Path):
+        if path != Path(marker.path.removeprefix("/")):
+            redirects[path] = marker.redirect_contents
+
     for version, item in sitemap.items():
         if version.startswith("latest"):  # TODO: check type of item instead
             continue
 
-        redirects[plugin.site_root / version] = item.default_marker.redirect_contents
+        add_redirect(item.default_marker, plugin.site_root / version)
         for lang, marker in item.markers.items():
-            redirects[plugin.site_root / version / lang] = marker.redirect_contents
+            add_redirect(marker, plugin.site_root / version / lang)
 
         item_version = Version(version)
         if not root_version or item_version > root_version: