+++ src/hexdoc/plugin/mod_plugin.py
@@ -14,6 +14,7 @@ from hexdoc.utils import ContextSource
 from .types import HookReturn
 
 if TYPE_CHECKING:
+    from hexdoc.cli.app import LoadedBookInfo
     from hexdoc.core import Properties
     from hexdoc.graphics.validators import ItemImage
 
@@ -139,6 +140,33 @@ class ModPlugin(ABC):
         rendering the book.
         """
 
+    def pre_render_site(
+        self,
+        props: Properties,
+        books: list[LoadedBookInfo],
+        env: SandboxedEnvironment,
+        output_dir: Path,
+    ) -> None:
+        """Called once, after hexdoc is done setting up the Jinja environment but before
+        rendering the book."""
+
+    def post_render_site(
+        self,
+        props: Properties,
+        books: list[LoadedBookInfo],
+        env: SandboxedEnvironment,
+        output_dir: Path,
+    ) -> None:
+        """Called once, at the end of `hexdoc build`."""
+
+    def pre_render_book(self, template_args: dict[str, Any], output_dir: Path) -> None:
+        """Called once per language, just before rendering the book for that
+        language."""
+
+    def post_render_book(self, template_args: dict[str, Any], output_dir: Path) -> None:
+        """Called once per language, after all book files for that language are
+        rendered."""
+
     def item_image_types(self) -> HookReturn[type[ItemImage[Any]]]:
         """List of `ModelImage` types to attempt when loading a block/item model."""
         return []
@@ -179,6 +207,14 @@ class ModPlugin(ABC):
         * value: `v/latest/main`
         """
         return self.site_root / "latest" / self.branch
+
+    @property
+    def flags(self) -> dict[str, bool]:
+        """Set default values for Patchouli config flags.
+
+        Unlike the table in `hexdoc.toml`, this is available for use in dependents.
+        """
+        return {}
 
 
 class VersionedModPlugin(ModPlugin):